Moved uninstall-specific classes into install package
